OXDEV-351 Deprecate getRevision() and related
Reason:
There is no such concept as a revision any more in OXID eShop v.6.0.0. and the file pkg.rev was removed from the code base a while ago. These methods will always return false now, as the underlying functionality was removed.

The methods which returned the revision of OXID eShop in version 5.3 were deprecated in various classes:
- \OxidEsales\EshopCommunity\Application\Model\Diagnostics::getRevision()
- \OxidEsales\EshopCommunity\Application\Model\FileChecker::getRevision()
- \OxidEsales\EshopCommunity\Core\Config::getRevision()
- \OxidEsales\EshopCommunity\Core\Controller\BaseController::getRevision()

Related class properties and setters were also removed:
- \OxidEsales\EshopCommunity\Application\Model\Diagnostics::$_sRevision
- \OxidEsales\EshopCommunity\Application\Model\Diagnostics::setRevision()
- \OxidEsales\EshopCommunity\Application\Model\FileChecker::$_sRevision
- \OxidEsales\EshopCommunity\Application\Model\FileChecker::setRevision()

The test for Config::getRevision() was removed.